+++ provider/aliyun/getfunction.go
@@ -15,13 +15,13 @@ func (m *Manager) GetFunction(funcName string) (*function.Meta, error) {
 		return nil, err
 	}
 	return &function.Meta{
-		FunctionName: *v.FunctionName,
-		MemorySize: int64(*v.MemorySize),
-		Timeout: *v.Timeout,
-		Description: *v.Description,
-		CreatedTime: *v.CreatedTime,
-		CodeChecksum: *v.CodeChecksum,
+		FunctionName:         *v.FunctionName,
+		MemorySize:           int64(*v.MemorySize),
+		Timeout:              int64(*v.Timeout),
+		Description:          *v.Description,
+		CreatedTime:          *v.CreatedTime,
+		CodeChecksum:         *v.CodeChecksum,
 		EnvironmentVariables: v.EnvironmentVariables,
-		Runtime: *v.Runtime,
+		Runtime:              *v.Runtime,
 	}, nil
 }
